:sparkles: Adding kwarg "magnify" to klip.pro and adi.pro
"magnify" allows for the option to scale left- and right-side data according to two separate plate scale solutions for each side. Defaults to no magnification as most of the difference in plate scale should be due to the specific distortion solutions chosen and not due to actual optical differences between the two sides of the telescope.